conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
blas=1.0=openblas
bzip2=1.0.8=h620ffc9_4
ca-certificates=2023.01.10=hca03da5_0
certifi=2022.12.7=py310hca03da5_0
charset-normalizer=3.0.1=pypi_0
contourpy=1.0.7=pypi_0
cycler=0.11.0=pypi_0
deepsnap=0.2.1=pypi_0
flit-core=3.6.0=pyhd3eb1b0_0
fonttools=4.38.0=pypi_0
idna=3.4=pypi_0
jinja2=3.1.2=pypi_0
joblib=1.2.0=pypi_0
kiwisolver=1.4.4=pypi_0
libcxx=14.0.6=h848a8c0_0
libffi=3.4.2=hca03da5_6
libgfortran=5.0.0=11_3_0_hca03da5_28
libgfortran5=11.3.0=h009349e_28
libopenblas=0.3.21=h269037a_0
littleutils=0.2.2=pypi_0
llvm-openmp=14.0.6=hc6e5704_0
markupsafe=2.1.2=pypi_0
matplotlib=3.6.3=pypi_0
ncurses=6.4=h313beb8_0
networkx=3.0=pypi_0
numpy=1.23.5=py310hb93e574_0
numpy-base=1.23.5=py310haf87e8b_0
ogb=1.3.5=pypi_0
openssl=1.1.1s=h1a28f6b_0
outdated=0.2.2=pypi_0
packaging=23.0=pypi_0
pandas=1.5.3=pypi_0
pillow=9.4.0=pypi_0
pip=22.3.1=py310hca03da5_0
psutil=5.9.4=pypi_0
pyparsing=3.0.9=pypi_0
python=3.10.9=hc0d8a6c_0
python-dateutil=2.8.2=pypi_0
pytorch=1.13.1=py3.10_0
pytz=2022.7.1=pypi_0
readline=8.2=h1a28f6b_0
requests=2.28.2=pypi_0
scikit-learn=1.2.1=pypi_0
scipy=1.10.0=pypi_0
setuptools=65.6.3=py310hca03da5_0
six=1.16.0=pypi_0
sqlite=3.40.1=h7a7dc30_0
threadpoolctl=3.1.0=pypi_0
tk=8.6.12=hb8d0fd4_0
torch-geometric=2.2.0=pypi_0
torch-scatter=2.1.0=pypi_0
torch-sparse=0.6.16=pypi_0
tqdm=4.64.1=pypi_0
typing_extensions=4.4.0=py310hca03da5_0
tzdata=2022g=h04d1e81_0
urllib3=1.26.14=pypi_0
wheel=0.37.1=pyhd3eb1b0_0
xz=5.2.10=h80987f9_1
zlib=1.2.13=h5a0b063_0
